requirements.txt
fastapi
uvicorn

pandas
numpy

scikit-learn>=1.4.0
xgboost

mlflow
shap

pydantic

matplotlib
seaborn

python-dotenv
